refactor(api): split monolithic router into domain sub-routers
Extract the 1,939-line router into a layered structure:

- lib/api/types.ts — shared types (Endpoint, ChartType, RankTopProfile,
  StatsRoutes, etc.) extracted from router.ts
- lib/api/routes/profiles.ts — profile/post endpoints, stats, search
- lib/api/routes/feed.ts — feed posts, trending, leaderboard
- lib/api/routes/wallet.ts — wallet activity, engagement (auth-gated)
- lib/api/routes/referral.ts — referral generate, redeem, genesis
- lib/api/routes/system.ts — charts, txs, vote activity, instance
- lib/api/router.ts — Parameters validators + API class shell, mounts
  sub-routers via the existing `routers` constructor param
- lib/api.ts — reduced from 1,939 lines to a single barrel re-export

Each sub-router is a factory `createXRouter(deps) -> Router`. Dependencies
(db, temporal, authCache, state) passed by closure. No DI, no abstract
base, no change to index.ts or the call site.

The deleted `lib/api/router.ts` (stale duplicate) now holds the
production code. import path in `lib/temporal/setup.ts` updated:
`../api` -> `../api/types`.
